Fix up Dockerfile and generation once again
