+++ Dockerfile
@@ -9,11 +9,12 @@ WORKDIR /scratch
 RUN go get github.com/go-bindata/go-bindata/...
 COPY . .
 COPY --from=frontend /scratch/dist frontend/dist
-RUN go generate pkg/servers/api/web/*
+RUN go generate cmd/talaria/*
 RUN CGO_ENABLED=0 go build -o talaria cmd/talaria/*
 
 FROM scratch
 ENV PATH /bin
 COPY --from=backend /scratch/talaria /bin/talaria
 EXPOSE 8080
+EXPOSE 8081
 ENTRYPOINT ["talaria", "server"]
